A great description of Tree Notation!
Hi Hari (@SRS-WRKS),

I am not making any substantial change to your file. GitHub does not make it easy to comment on the contents of a file, so this is my attempt! The couple of lines describing Tree Notation is the best explanation that I have seen on the topic. I'll have to go back and look at some previous notes that I've made, but this certainly has directed me on the right path! 

Thanks for the explanation!

Kind Regards,
Liam